GH Actions: version update for `actions/checkout`
V3 runs on Node 16, which is deprecated for action runners. v4 runs on Node 20 which should fix this issue.

This fixes the following notices in GH Actions Summaries:
```
The following actions use a deprecated Node.js version and will be forced to run on node20: actions/checkout@v3. For more info: https://github.blog/changelog/2024-03-07-github-actions-all-actions-will-run-on-node20-instead-of-node16-by-default/
```

Ref: https://github.com/actions/checkout/releases