PHP 8.4 | CallErrorException: prevent deprecation for `E_STRICT` [2]
The `E_STRICT` constant is deprecated as of PHP 8.4 and will be removed in PHP 9.0.

The error level hasn't been in use since PHP 8.0 anyway and was only barely still used in PHP 7.x.

All the same as this is an error handling class, I'm proposing to still handle it, but conditionally to prevent the deprecation notice.

The solution applied is inspired by a similar fix made in PHPUnit by @sebastianbergmann.

Ref:
* https://wiki.php.net/rfc/deprecations_php_8_4#remove_e_strict_error_level_and_deprecate_e_strict_constant